Server can now communicate with API. #1
